Add description properry to fields and create editable text
