+++ app/src/features/ApplicationWorkspace/constants.ts
@@ -1,28 +1,11 @@
 import {faker} from '@faker-js/faker';
 
-const fieldTypes = [
-  'short_text',
-  'long_text',
-  'number',
-  'email',
-  'dropdown',
-  'multiple_choice',
-  'yes_no',
-] as const;
-
-type FieldTypeTuple = typeof fieldTypes;
-
-export type FieldTypes = FieldTypeTuple[number];
-
-export interface Field {
-  id: string;
-  title: string;
-  type: FieldTypes;
-}
+import {fieldTypes} from './workspaceSlice';
 
 export const fieldBuilder = () => ({
   id: faker.string.numeric(4),
   title: faker.lorem.sentence(),
+  description: faker.lorem.sentence(),
   type: faker.helpers.arrayElement(fieldTypes),
 });
